+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Documentation improved
 
+### Added
+- `add` method for DirFS (adds an existing file or directory)
+
 ## [0.1.10] - 2026-02-04
 
 ### Fixed